+++ bia-shared-datamodels/src/bia_shared_datamodels/package_specific_uuid_creation/biostudies_ingest_uuid_creation.py
@@ -2,7 +2,8 @@ from uuid import UUID
 from bia_shared_datamodels import uuid_creation, attribute_models
 from bia_shared_datamodels.semantic_models import Provenance
 from typing import Optional
-from .shared import create_unique_str_attribute
+from bia_shared_datamodels.package_specific_uuid_creation import shared
+
 
 def create_dataset_uuid(
     study_uuid: str, biostudies_section_accno: str
@@ -12,7 +13,7 @@ def create_dataset_uuid(
         uuid_creation.create_dataset_uuid(
             study_uuid=study_uuid, unique_string=unique_string
         ),
-        create_unique_str_attribute(unique_string, Provenance.bia_ingest),
+        shared.create_unique_str_attribute(unique_string, Provenance.bia_ingest),
     )
 
 
@@ -28,7 +29,7 @@ def create_bio_sample_uuid(
         uuid_creation.create_bio_sample_uuid(
             study_uuid=study_uuid, unique_string=unique_string
         ),
-        create_unique_str_attribute(unique_string, Provenance.bia_ingest),
+        shared.create_unique_str_attribute(unique_string, Provenance.bia_ingest),
     )
 
 
@@ -40,7 +41,7 @@ def create_protocol_uuid(
         uuid_creation.create_protocol_uuid(
             study_uuid=study_uuid, unique_string=unique_string
         ),
-        create_unique_str_attribute(unique_string, Provenance.bia_ingest),
+        shared.create_unique_str_attribute(unique_string, Provenance.bia_ingest),
     )
 
 
@@ -52,7 +53,7 @@ def create_annotation_method_uuid(
         uuid_creation.create_annotation_method_uuid(
             study_uuid=study_uuid, unique_string=unique_string
         ),
-        create_unique_str_attribute(unique_string, Provenance.bia_ingest),
+        shared.create_unique_str_attribute(unique_string, Provenance.bia_ingest),
     )
 
 
@@ -65,7 +66,7 @@ def create_specimen_imaging_preparation_protocol_uuid(
         uuid_creation.create_specimen_imaging_preparation_protocol_uuid(
             study_uuid=study_uuid, unique_string=f"{biostudies_section_accno}"
         ),
-        create_unique_str_attribute(unique_string, Provenance.bia_ingest),
+        shared.create_unique_str_attribute(unique_string, Provenance.bia_ingest),
     )
 
 
@@ -77,5 +78,5 @@ def create_image_acquisition_protocol_uuid(
         uuid_creation.create_image_acquisition_protocol_uuid(
             study_uuid=study_uuid, unique_string=unique_string
         ),
-        create_unique_str_attribute(unique_string, Provenance.bia_ingest),
+        shared.create_unique_str_attribute(unique_string, Provenance.bia_ingest),
     )
