+++ bia-shared-datamodels/src/bia_shared_datamodels/package_specific_uuid_creation/ro_crate_uuid_creation.py
@@ -1,7 +1,7 @@
 from uuid import UUID
 from bia_shared_datamodels import uuid_creation, attribute_models
 from bia_shared_datamodels.semantic_models import Provenance
-from .shared import create_unique_str_attribute
+from bia_shared_datamodels.package_specific_uuid_creation import shared
 
 
 def create_dataset_uuid(
@@ -12,7 +12,7 @@ def create_dataset_uuid(
         uuid_creation.create_dataset_uuid(
             study_uuid=study_uuid, unique_string=unique_string
         ),
-        create_unique_str_attribute(unique_string, Provenance.bia_ingest),
+        shared.create_unique_str_attribute(unique_string, Provenance.bia_ingest),
     )
 
 
@@ -24,7 +24,7 @@ def create_bio_sample_uuid(
         uuid_creation.create_bio_sample_uuid(
             study_uuid=study_uuid, unique_string=unique_string
         ),
-        create_unique_str_attribute(unique_string, Provenance.bia_ingest),
+        shared.create_unique_str_attribute(unique_string, Provenance.bia_ingest),
     )
 
 
@@ -36,7 +36,7 @@ def create_protocol_uuid(
         uuid_creation.create_protocol_uuid(
             study_uuid=study_uuid, unique_string=unique_string
         ),
-        create_unique_str_attribute(unique_string, Provenance.bia_ingest),
+        shared.create_unique_str_attribute(unique_string, Provenance.bia_ingest),
     )
 
 
@@ -48,7 +48,7 @@ def create_annotation_method_uuid(
         uuid_creation.create_annotation_method_uuid(
             study_uuid=study_uuid, unique_string=unique_string
         ),
-        create_unique_str_attribute(unique_string, Provenance.bia_ingest),
+        shared.create_unique_str_attribute(unique_string, Provenance.bia_ingest),
     )
 
 
@@ -60,7 +60,7 @@ def create_specimen_imaging_preparation_protocol_uuid(
         uuid_creation.create_specimen_imaging_preparation_protocol_uuid(
             study_uuid=study_uuid, unique_string=f"{ro_crate_id}"
         ),
-        create_unique_str_attribute(unique_string, Provenance.bia_ingest),
+        shared.create_unique_str_attribute(unique_string, Provenance.bia_ingest),
     )
 
 
@@ -72,5 +72,5 @@ def create_image_acquisition_protocol_uuid(
         uuid_creation.create_image_acquisition_protocol_uuid(
             study_uuid=study_uuid, unique_string=unique_string
         ),
-        create_unique_str_attribute(unique_string, Provenance.bia_ingest),
+        shared.create_unique_str_attribute(unique_string, Provenance.bia_ingest),
     )
